test: proper placement and keeping old ones
